requirements.txt
certifi==2018.4.16
chardet==3.0.4
click==6.7
Django==2.0.6
feedparser==5.2.1
Flask==1.0.2
httpsig==1.2.0
idna==2.6
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
numpy==1.14.4
pandas==0.23.0
Pillow==5.1.0
pkg-resources==0.0.0
pycrypto==2.6.1
pycryptodome==3.4.7
python-dateutil==2.7.3
python-magic==0.4.15
pytz==2018.4
requests==2.19.0
six==1.11.0
urllib3==1.22
uWSGI==2.0.17
Werkzeug==0.14.1
